@@ -1,27 +1,42 @@
 {
-  "name": "eslint-plugin-bedrock",
-  "version": "1.0.27",
+  "name": "@bedrockio/eslint-plugin",
+  "version": "1.0.0",
   "description": "Common ESLint plugin for Bedrock projects.",
   "license": "MIT",
+  "type": "module",
   "keywords": [
     "eslint",
     "eslintplugin",
     "eslint-plugin"
   ],
   "main": "index.js",
+  "dependencies": {
+    "@eslint/compat": "^1.0.1",
+    "@eslint/js": "^9.2.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jest": "^28.5.0",
+    "eslint-plugin-mdx": "^3.1.5",
+    "eslint-plugin-react": "^7.34.1",
+    "globals": "^15.2.0"
+  },
   "peerDependencies": {
-    "eslint": ">=8",
-    "eslint-import-resolver-webpack": "^0.13.2"
+    "eslint": ">=9",
+    "eslint-import-resolver-webpack": "^0.13.8"
   },
   "peerDependenciesMeta": {
     "eslint-import-resolver-webpack": {
       "optional": true
+    },
+    "eslint-plugin-react": {
+      "optional": true
+    },
+    "eslint-plugin-mdx": {
+      "optional": true
     }
   },
-  "dependencies": {
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jest": "^28.5.0",
-    "eslint-plugin-react": "^7.34.1"
+  "devDependencies": {
+    "eslint": "^9.2.0",
+    "eslint-import-resolver-webpack": "^0.13.8"
   },
   "volta": {
     "node": "20.12.2",
