@@ -1,10 +1,19 @@
-module.exports = {
+import { default as recommended } from './recommended.js';
+import { default as react } from './react.js';
+import { default as jest } from './jest.js';
+import { default as mdx } from './mdx.js';
+import { default as nodeImports } from './node-imports.js';
+import { default as webpackImports } from './webpack-imports.js';
+import { default as typescriptImports } from './typescript-imports.js';
+
+export default {
   configs: {
-    recommended: require('./recommended.json'),
-    'imports-node': require('./imports-node.json'),
-    'imports-webpack': require('./imports-webpack.json'),
-    'imports-typescript': require('./imports-typescript.json'),
-    react: require('./react.json'),
-    jest: require('./jest.json'),
+    mdx,
+    jest,
+    react,
+    recommended,
+    nodeImports,
+    webpackImports,
+    typescriptImports,
   },
 };
